USDC tests working with funds, loans, sales tests
