+++ package.json
@@ -5,7 +5,7 @@
     "@liquality/crypto": "^0.2.10",
     "@liquality/ethereum-utils": "^0.2.10",
     "@mblackmblack/to-seconds": "^0.1.0",
-    "axios": "^0.19.0",
+    "axios": "^0.19.0", 
     "bignumber.js": "^9.0.0",
     "decimal.js": "^10.2.0",
     "lodash": "^4.17.13",
@@ -17,21 +17,16 @@
   },
   "scripts": {
     "test": "truffle test",
-    "test:dai:funds": "truffle test ./test/daiFunds.js",
-    "test:dai:loans": "truffle test ./test/daiLoans.js",
-    "test:dai:sales": "truffle test ./test/daiSales.js",
-    "test:dai": "truffle test ./test/daiFunds.js ./test/daiLoans.js ./test/daiSales.js",
+    "test:funds": "truffle test ./test/daiFunds.js",
+    "test:loans": "truffle test ./test/daiLoans.js",
+    "test:sales": "truffle test ./test/daiSales.js",
     "test:compound": "truffle test ./test/compound.js",
     "test:alcompound:js": "truffle test ./test/alcompound.js",
     "test:alcompound:sol": "truffle test ./test/TestALCompound.sol",
     "test:alcompound": "truffle test ./test/alcompound.js ./test/TestALCompound.sol",
     "test:e2e": "truffle test ./test/e2e.js",
-    "test:global_interest_increase": "truffle test ./test/globalInterestIncrease.js",
-    "test:global_interest_decrease": "truffle test ./test/globalInterestDecrease.js",
-    "test:usdc:funds": "truffle test ./test/usdcFunds.js",
-    "test:usdc:loans": "truffle test ./test/usdcLoans.js",
-    "test:usdc:sales": "truffle test ./test/usdcSales.js",
-    "test:usdc": "truffle test ./test/usdcFunds.js ./test/usdcLoans.js ./test/usdcSales.js",
+    "test:interest_increase": "truffle test ./test/interestIncrease.js",
+    "test:interest_decrease": "truffle test ./test/interestDecrease.js",
     "coverage": "npx solidity-coverage"
   },
   "devDependencies": {
